backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
plaid-python==35.0.0
pydantic==2.5.0
httpx==0.25.2
python-multipart==0.0.6
requests==2.31.0
